Allow CLUSTER_NAME and COMMON_DOMAIN to be given in GuestConfig (#85)
There are use cases where multiple guest clusters need to be created
concurrently and use of environment variables for CLUSTER_NAME is
slightly difficult to safely manage. Therefore allow it to be specified
in GuestConfig. To align with this, also COMMON_DOMAIN can be defined
via config variable. When either of given fields is empty, this
implementation defaults to environment variable in order to maintain
current functionality.