framework: improve latest/previous version inference (previous wip/current)
After chat with @xh3b4sd we decided our current nomenclature for wip and
current version doesn't make much sense. Especially in the advent of
versioning of operators. Also what we want to do is basically always
test two latest versions and upgrades from the previous to the latest.
It would be cool to do more upgrade checks but it takes resources and
this thing here seems to be reasonable choice.