circleci: disable e2eRemoteTest
Towards https://github.com/giantswarm/giantswarm/issues/4648

This test is most likely related to shipyard. This is a component to run
remote e2e control planes in AWS IIRC. Tests stopped to work out of the
sudden. We need to investigate. I will disable those tests for now.

We also need to consider if we still need shipyard at all. It was
abandoned for a while since we use localkube. But it still may come in
handy if we need to write vault encrypter tests in aws-operator.

Example build: https://circleci.com/gh/giantswarm/e2e-harness/2179