Increase release.waitForConditions timeout to 10m (#175)
release.waitForConditions is used e.g. for ensuring that API secret gets
deleted after corresponding chart is deleted. Occasionally it seems that
it takes moderately long time for cert-operator to get execution time
and pick up the delete event so we give some room for it to operate by
increasing wait timeout value.