Fix minikube RBAC mode configuration instruction (#200)
Since minikube 26.0, kubeadm is the default bootstrapper for minikube, localkube has been deprecated.
README.md instruction how to start minikube with RBAC enabled was using localkube configuration style which is not compatible with kubeadm; it would cause minikube startup to hang.
This PR fixes the instruction how to start minikube with RBAC enabled to be compatible with kubeadm i.e. recent minikube versions.